Fix infinite loop bug.
Cause: an on_commit callback creates a transaction, which afterwards loops through all the on_commit records waiting to be called, which includes the one that creates a transaction.

Fix: Have a commit object stack that gets pushed and popped for each commit object, so that it doesn't loop, and (untested) its inner-transaction's on_commit can run.